refactor(muya): split TextSelection and turn Selection into a facade
